+++ CHANGES.rst
@@ -1,6 +1,9 @@
 0.2 (in development)
 ====================
 
+- Merged ``paste-new-from-buffer`` and ``paste-new-from-region`` into a new,
+  more convenient ``paste-new`` function.
+
 
 0.1.2 (Aug 03, 2010)
 ====================